Updating externals from utils svn to most recent revision. Updating Vizmo to use the new mathtool step 1: which is making vizmo use the new quaternion, vector, and angle conversions appropriately. Most changes are name changes of Vector3D to Vector3d and Quaternion conversion directly to euler angles.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@356 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ src/GUI/CaptureOptions.cpp
@@ -119,7 +119,7 @@ CaptureOptions::CaptureMovie(){
   //Pop up a MovieSaveDialog
   MovieSaveDialog msd(this, Qt::Dialog); 
   if(msd.exec() == QDialog::Accepted){ 
-    size_t digits = max(msd.m_frameDigits, log10(msd.m_endFrame/msd.m_stepSize) + 2);
+    size_t digits = max(double(msd.m_frameDigits), log10(msd.m_endFrame/msd.m_stepSize) + 2);
 
     //Create the progress bar for saving images
     QProgressDialog progress("Saving images...", "Abort", msd.m_startFrame, msd.m_endFrame, this);
